:recycle: refactor: move some objects to vo
