+++ blog-api/src/main/java/top/naccl/model/vo/Copyright.java
@@ -1,4 +1,4 @@
-package top.naccl.model.bean;
+package top.naccl.model.vo;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -6,7 +6,7 @@ import lombok.Setter;
 import lombok.ToString;
 
 /**
- * @Description: 自定义爱好
+ * @Description: copyright
  * @Author: Naccl
  * @Date: 2020-08-09
  */
@@ -14,7 +14,7 @@ import lombok.ToString;
 @Getter
 @Setter
 @ToString
-public class Favorite {
+public class Copyright {
 	private String title;
-	private String content;
+	private String siteName;
 }
