+++ blog-api/src/main/java/top/naccl/service/impl/SiteSettingServiceImpl.java
@@ -7,9 +7,9 @@ import top.naccl.constant.RedisKeyConstants;
 import top.naccl.entity.SiteSetting;
 import top.naccl.exception.PersistenceException;
 import top.naccl.mapper.SiteSettingMapper;
-import top.naccl.model.bean.Badge;
-import top.naccl.model.bean.Copyright;
-import top.naccl.model.bean.Favorite;
+import top.naccl.model.vo.Badge;
+import top.naccl.model.vo.Copyright;
+import top.naccl.model.vo.Favorite;
 import top.naccl.model.vo.Introduction;
 import top.naccl.service.RedisService;
 import top.naccl.service.SiteSettingService;
